+++ src/maps/tiles.rs
@@ -45,10 +45,10 @@ impl<T: Tile> TileSet<T> {
         let rows_count = self.tiles.len() as i32;
         let columns_count = self.tiles[0].len() as i32;
 
-        let min_y = viewport.y as i32 - 2;
-        let max_y = (viewport.y + viewport.h) as i32 + 4;
-        let min_x = viewport.x as i32 - 2;
-        let max_x = (viewport.x + viewport.w) as i32 + 4;
+        let min_y = viewport.y - 2;
+        let max_y = (viewport.y + viewport.h) + 4;
+        let min_x = viewport.x - 2;
+        let max_x = (viewport.x + viewport.w) + 4;
 
         let min_row = min_y.max(0).min(rows_count) as usize;
         let max_row = max_y.max(0).min(rows_count) as usize;
